apps/workers/requirements.txt
# Created automatically by Cursor AI (2025-01-27)

# Core dependencies
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database
asyncpg>=0.29.0,<1.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0

# PDF processing
pdfminer.six>=20221105
pymupdf>=1.23.0,<2.0.0
pytesseract>=0.3.10
Pillow>=10.0.0,<11.0.0

# OCR
easyocr>=1.7.0,<2.0.0

# AI/ML
openai>=1.3.0,<2.0.0
langchain>=0.1.0,<1.0.0
langchain-openai>=0.0.2,<1.0.0
langchain-text-splitters>=0.0.1,<1.0.0
numpy>=1.24.0,<2.0.0

# Vector databases
pgvector>=0.2.0,<1.0.0
faiss-cpu>=1.7.4,<2.0.0

# Message queue
nats-py>=2.3.0,<3.0.0
redis>=5.0.0,<6.0.0

# Storage
boto3>=1.34.0,<2.0.0
minio>=7.2.0,<8.0.0

# Virus scanning
clamd>=1.0.2,<2.0.0

# Utilities
python-multipart>=0.0.6,<1.0.0
httpx>=0.25.0,<1.0.0
tenacity>=8.2.0,<9.0.0
structlog>=23.2.0,<24.0.0
python-json-logger>=2.0.7,<3.0.0

# Development
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
mypy>=1.7.0,<2.0.0
ruff>=0.1.0,<1.0.0
